STY:toto:matchrule_validators: MATCH variable rename
The MATCH matchrule handler had two variables, target and new_target.
These names didn't match the ones for the fields in the specification or
anywhere else in the code. Rename these variables to match the
specification and the rest of the codebase (target -> path_pattern,
new_target -> target_path_pattern).