Replaces command alignment exception with warning
In a previous commit we added a custom CommandAlignmentFailed
exception and replaced log with print in case of soft fail.
This commit removes the custom exception (1) and changes print back
to log (2) for the following reasons.

(1) We decided use log throughout the library also for user feedback
because it is more powerful and can easily be toggled
programatically, see discussion
https://github.com/in-toto/in-toto/issues/6#issuecomment-263988714

(2) We decided to never hard fail command alignment because the check
is of informational character rather than a security guarantee. See
https://github.com/in-toto/in-toto/issues/46

The commit also adds unit tests for command alignment