Adds apply exclude patterns to mid-level dirs
Exclude patterns need to be applied three times:
(1) On the first level of the directory tree when artifacts could
be either files or directories,
(2) on intermediate directories, i.e. dirs that are not on the first
level, and
(3) on tree leaves that are not on the first level.

EXAMPLE:
.
├── foo.pyc
└── sub
    ├── .git
    └── bar.pyc
ARTIFACT_EXCLUDES = [".git", "*.pyc]
record_artifacts_as_dict(".")

(1) excludes foo.pyc
(2) excludes sub/.git
(3) excludes bar.pyc

This commit adds (2) + unit tests