Adds a same-hash artifact to RecordArtifacts test
Hash collisions should not affect the way artifacts are recorded.
They are only relevant when mapping files from one artifact
dict to another, e.g. while using match rules.

We add two artifacts with the same hash to the record artifacts
TestCase to demonstrate that it does not have an effect.