IDs not in link files...
I didn't find the IDs in the link files so had to change the directions to accommodate.  

Discard this PR if this is because my in-toto software is out of date.  (However, if this is so perhaps change the directions to use * in the cp commands in place of the keyIDs...