Fixes dead symlink bug in `record_artifacts_as_dict`
File existence is only checked on the first level of the file
system traversal. If we recurse in a directory using os.walk,
we might also record symlinks to non-existent files, which would
crash once we try reading the file.

This commit excludes non-existent files in our `os.walk` loop
and adapts the related unittest.