Modifies expected_command field
The expected_command field in steps of a layout are now stored
as a list, simliar to the command field of link, against which
the expected_command is compared.

Expected commands can still be fed to the `Step.read` function
as String, but they are split there using the shlex package
to allow shell like (escaping) syntax.