Moves default assignments to contructors
The in-toto metadata classes used to perform some default attribute
assignments in their static `read` methods i.e. methods
that create the respective object of the class from a passed
dictionary using the auto-constructors from the `attr` module. This
lead to inconsistencies.

This commit adds a constructor to `Signable`, `Layout`, `Step`,
`Inspection` and `Link` and moves the default attribute logic
from the read method to the constructor (called by the read method).

Note: The constructors henceforth only accept keyword arguments.