Update _type field in test link metadata files
This commit updates the _type field ("Link" -> "link") in the
two demo/test link metadata files and re-signs them with the
respective demo/test keys "bob" (write-code.c8650e01.link) and
"carl" (package.2dc02526.link).

Note:
While updating the signatures I realized that the keyids,
as they appear e.g. in the link filenames and in
"demo.layout.template", are actually outdated. I.e. if
these keys ("bob" and "carl") are imported from PEM format
into securesystemlib's rsa key format and hashed to generate
the keyid, that keyid would be different.
The reason is that securesystemlib's key format has changed
to to include a field "keyid_hash_algorithms" when
creating the hash/keyid.

This should be taken care of in in-toto/in-toto#87