Homogenize testing demo keys
The meta data files in 'test/demo_files' were signed by keys
that are not longer in the repository. Neither the key pairs
for alice, bob and carl in 'test/demo_files', nor the key pairs
for alice, bob and carl in the demo directory were used.
Moreover the testing key pairs and the demo key pairs do not
correspond.

While this does not break unittesting, which uses the keys and
the meta data files independently. It is confusing and leads to
problems if the existing test meta data needs to be resigned,
e.g. because of changes in the meta data format.

This commit, as first part of a fix, homogenizes the key pairs
in test/demo_files and the demo directory for alice, bob and
carl respectively