ENH: in-toto/models: update structure to match spec
The models were restructured to match the specification. Specifically,
the fields were separated into two groups:

- Signed: the fields from which the signature was computed (or is to be
computed)
- Signatures: The signatures over the signed fields.

To do this, I introduced a custom __getattr__ module to reduce the size
of the patch and make most of the tests pass. This is not the behavior
that we want in the long run, but it will let us slowly migrate to a
solid solution.