ENH: in-toto/runlib: support new format
runlib instantiates link metadata in many of its methods. Due to the
nature of the initializer, the fields of the metadata are not set
properly: they are not set in the nested 'signed' property, but
"besides" it. This could've been fixed with a custom setattr for the
Metablock class, but that solution was its own can of worms.