Remove mock_link and simplify in_toto_mock
The MockLink model is no longer needed because its only addition
to regular Links used to be a working_dir field, which is now
also stored as part of a regular Links' environment field.

This commit also simplifies `runlib.in_toto_mock` to wrap
`runlib.in_toto_run` instead of replicating it.