Remove implicit `DISALLOW *` from rule verification
Until now every recorded artifact had to be consumed by a rule.
That is, when verifying the batch of rules for a certain artifact
type (material or product) reported for a step or inspection, the
verification routine would keep a temporary queue of all
artifacts of that type for that step or inspection and remove
them from the queue each time an artifact was matched by a rule's
artifact pattern.

If any artifacts remained in the queue after verifying all rules
(for that artifact type of that step or inspection) a
RuleVerficationError was raised.

This behavior amounted to an implicit `DISALLOW *` at the end of
a given `expected_materials` or `expected_products` list.

With this change the implicit `DISALLOW *` is removed, i.e. replaced
with an implicit `ALLOW *`.

However a layout designer can still add an explicit `DISALLOW *`
at the end of a given rule list to fail if corresponding
artifacts are not matched by any prior rule's pattern.

The rationale for this change can be found in in-toto/docs#4.