Reorder final product verification routines
Postpone Inspection verification routine to after having
verified expected artifacts of Steps.
This guarantees that Inspection commands don't run on
compromised target files, which would be a surface for attacks.

As a consequence MATCH rules defined for Steps won't be able
to reference Inspection artifacts, because the required Inspection
link metadata only gets generated when running the Inspections.

However, MATCH rules defined for Inspection are still able
to reference Step artifacts.

This commit also updates and fixes a couple of minor issues in the
long docstring of `verifylib.in_toto_verify`, which describes the
final product verification routine.