Move Metablock class to own module
This commit does not change any logic, it simply moves the
Metablock class from the common module to a new metadata module
and updates the affected import statements.

In the future Metablock won't be an abstract superclass for Link
or Layout (adapters) but instead will be a concrete metadata
class that contains instances of Link or Layout.

For clarity and to avoid problems with circular imports (Metablock
will require Link and Layout, and Link and Layout will require
ValidationMixin from the common module, ...) this commit moves the
Metablock class from the common module to the metadata module.