Move and update link loading method (WIP)
Note: This is part of a series of refactoring commits.

This commit moves the function that loads all the links as defined
in a layout from a static method of the Layout class to an
autonomous function in verifylib and updates the function to
fit the WIP adapter refactoring.

Having this function as class method in Layout does not seem
intuitive anymore, the function should rather be seen as part of
the whole final product verification process. This commit also
helps if we we want to further reduce OOP concepts in the future.