Remove Link and Layout dump methods (WIP)
Note: This is part of a series of refactoring commits.

This commit removes the methods to write a Link or Layout object
to disk.

In the future Link and Layout objects should only be written to
and read from disk when contained in a Metablock object
(add signatures). Hence, Metablock provides the loading and
dumping methods.