Move read Link/Layout-from-dict (WIP)
Note: This is part of a series of refactoring commits.

This commit moves the static read methods that are used to parse
a dictionary in order to create a Link or Layout object to a
place where they will be preserved, since in the course of this
refactor the adapter classes will be removed.