Purge adaptor pattern (WIP)
Note: This is part of a series of refactoring commits.

This commit fully removes the adaptor pattern used to model the
relationship between generic signed Metadata and Link and Layout
objects.

The pattern was introduced to facilitate the transation
from *Link or Layout `is-a` signed metadata* to *signed metadata
`has-a` Link or layout* relationship cf.i in-toto/in-toto#120

Unfortunately, the pattern added a lot of extra complexity, which
made the code harder to understand and as a consequence introduced
bugs.

The commit removes all adaptor related logic from Metadata, Layout
and Link classes.

Furthermore it removes the adaptor classes Link and Layout and
renames the data holding classes LinkSignable and
LayoutSignable back to Link and Layout respectively.