Add module to read envvars and rcfiles
This module provides methods to read environment variables and
rcfiles and to write them to the in_toto.settings module, i.e.
override them.

The commit also adds a call of the settings parsing and overriding
method to in-toto's __init__.py, which should be the first thing
that's executed when using in_toto. Any code that accesses
in_toto.settings thereafter receives the values defined by the
user.