Move user settings from __init__ to cli tools
As pointed out by @vladimir-v-diaz it is odd to perform actions
like monkeypatching a module using env variables and rcfiles
when importing a package or module.

This behavior is better suited when the package is used via
command line tools. Hence, this commit removes the a call to
`user_settings.set_settings` from __init__ and instead calls the
function in the command line interfaces.