Bump release v0.1.1
It turns out that PyPi's description parser breaks if there is
a newline in the short description string.
As a consequence it does not display the string in
`long_description` but a "summary" of the fields passed to setup.

This commit removes the newlines from description and
long description, bumps the maintenance version number and
replaces 0.1.0 with 0.1.1 (we can't replace packages on pypi but
we can discard old ones).