Re-add gpg signing to in_toto_run and runlib
Adds new command line arguments to in-toto-run and arguments to
runlib.in_toto_run to alternatively sign a resulting link metadata
file with a gpg key identified by the passed gpg key id or with the
default gpg key in the keyring. Additionally a path to a gpg
keyring can be passed to not use the default gpg keyring.

This commit also adds a sign_gpg method to Metablock that uses the
new in_toto.gpg module instead of securesystemslib.

This commit also modifies both Metablock.sign and
Metablock.sign_gpg to return the newly created signature. This is
useful to e.g. obtain the used keyid, which is not available
beforehand when signing with the GPG default key.