Add workaround if gpg signing can't compute keyid
This commit adds a workaround to gpg_sign_object if the created
signature is returned with an empty keyid.
The workaround consists of exporting the public part of the signing
key to compute the keyid.
Note: this will fail if gpg_sign_object is called without a keyid
(i.e. using the default key), because exporting the public key
requires a keyid.