Use securesystemslib.encode_canonical
This commit removes the external dependency of `canonicaljson`
and instead uses `securesystemslib.encode_canonical` to sign
and verify the signable part of in_toto metadata.

The commit also adds a new property to the Signable class which
returns that canonical json string (used to be returned by `repr`),
to be more explicit about what is used for signing/verification.

Metablock's and Signable's `repr` methods now return a
non-canonical indented JSON string. Former is used to write
metadata to disk. Metadata does not have to be displayed
(represented) or stored in canonical json, this is only required
to generated and verify signatures.

The commit also updates the signatures of the metadata test files.