Comment out unused user settings
Settings are currently only available for in-toto-run/record.
There is no need to parse user settings from envvars and rcfiles
in other in-toto command line interfaces, this just pollutes the
log (confusing the user) with lines like:

```
Setting (default): ARTIFACT_EXCLUDE_PATTERNS=['*.link*', '.git', '*.pyc', '*~']
Setting (default): ARTIFACT_BASE_PATH=None
```

We can re-add the call to the user settings parser once we provide
settings for the tools in question as well.

We can re-think the (user) settings verbosity in general, e.g.
with in-toto/in-toto#117