Add gpg support to in-toto-record cli tool and lib
Note that for runlib.in_toto_record_stop we have to apply a
workaround if we want to use the default gpg key to be able
to load and verify the preliminary link file created with
in_toto_record_start, because the keyid required to load and
verify the right file is not available beforehand.
As a workaround we glob for any preliminary files (and fail if we
don't find exactly one) and then use the keyid from the files
signature.