Verify verifylib link signature verification
This commit combines 'verify_link_signatures' and
'verify_all_steps_signatures' into the function
'verify_link_signature_thresholds', which verifies signatures of
the passed links according to the layouts requirements (thresholds)
and authorizations (pubkeys).

If does not immediately fail if a passed link is not authorized or
has an invalid signature, as long as there are enough (threshold)
valid and authorized links in the end.

In the latter case the function returns an updated link dictionary
containing only the valid and authorized links.