Remove commented out user setting parsing
We commented out unused user settings parsing, which broke the
build because the linter complained about unused imports.
This commit removes the commented out calls and the imports
altogether.

The parser can be re-added to the modified command line scripts,
if we add user settings that are available in those scripts.