Fix packages section in setup.py
Despite all possible wildcard and "tests" dir combinations we
didn't exclude the only test directory that exists in the project.

This commit switches to explicitly naming the few packages we
want to include, which also obsoletes the `find_package` module.
If we want to go back to excluding unwanted packages we'd have
to use (already taking into account the soon to be merged, but
to be excluded "debian" dir):

packages=find_packages(exclude=["test", "test.*", "debian"])
or
packages=find_packages(exclude=["test", "test.models", "debian"])