Add pack_rule functions to artifact_rules
Steps and Inspections carry artifact rules in list forms. To
parse a rule in list form into a dictionary, which is easier
to work with, we use unpack_rule.

This commit adds a pack_rule function to create the rules based
on the fields in the rule_data function and a shortcut to
directly use the rule_data dictionary returned from unpack_rule
to get back to the rule in list format, e.g.:
pack_rule_data(unpack_rule(rule)).

Additionally this commit adds shortcuts for create, delete, modify,
allow and disallow rules.

This functions can be used for layout creation.