Add Step and Inspection superclass SupplyChainItem
The Step and Inspection class both carry expected_materials
and expected_products and already treated similarly during
final product verification, in terms of rule verification and.

Additionally both items have a name that is used to associate
link metadata.

This commit adds superclass SupplyChainItem (for the lack of a
better name) in order to reduce code redundancy.