FIX: gpg: s/pgp+SHA1/pgp+SHA2/
The code used to identify the pgp-compliant algorithms as PGP+sha1. In
reality, we are using sha256 as the hash algorithm. Change the default
string to pgp+SHA2 to properly document what are the methods used for
hashing.

In addition, the string has been moved to a constant within
gpg.formats.py so it can be easily replaced in the future.