Add IndexError guard in get_summary_link
get_summary_link creates a summary link for in entire layout
based on the links corresponding to the first and last step of
the layout.
This commit adds a guard to take care of the unlikely case
that a layout without steps is verified (plus a unittest for that
case).