Add link dir argument to in_toto_verify
Until now links corresponding to the steps in a layout to be
verified were loaded from the current working directory upon
verification.

This commit adds an optional link directory argument to
in_toto_verify and the subroutine load_links_for_layout to load
links from a passed path.

This change is required for sublayout link namespacing introduced
in a subsequent commit.