Add subkeys on gpg pubkey export
GPG returns the entire pubkey bundle (main key plus subkeys) when
any main or subkey keyid of that bundle is passed.  This commit
modifies our custom export pubkey function to also return the
entire pubkey bundle, returned by gpg and not only the main pubkey,
but only if the passed keyid matches the main key's keyid or any of
the subkeys' keyids.

Moreover, if a subkey keyid was passed, the user is warned, since
s/he might be unintentally authorizing the mainkey or a sibling
subkey.

The commit includes
- moving the parse_pubkeys_from_packets implementation to
  gpg_export_pubkey.
- creating the key dictionary in our custom format in
  parse_pubkey_packet instead of gpg_export_key (this is
  consistent with the function parse_signature_packet, which also
  returns a signature dict in our custom format)
- Fixes a unit test, that now fails with the newly added
  KeyNotFoundError