Add gpg parse_pubkey_payload function
This function is mostly identical with parse_pubkey_packet
(which it will replace in a subsequent commit) except that it
takes the payload of a pubkey packet instead of the whole packet.

The payload can be obtained by using
in_toto.gpg.util.parse_packet_header

This function is useful for a function that decides whether to
parse the payload of the packet based on info from the header.