Return short keyid from parse_signature_packet
So far parse_signature_packet only returned the full keyid
as taken from subpacket 33, which is not part of rfc4880 but an
experimental addition only available in newer versions of gpg
https://archive.cert.uni-stuttgart.de/openpgp/2016/06/msg00004.html

This commit also parses the short keyid from subpacket 16
(see of rfc4480 5.2.3.1) and returns it as additional optional
field in the signature dictionary.

The commit also updates the parse_subpackets function to return
the payload without its first byte, which describes the subpacket
type and is already returned separately, i.e.:
i.e. [(<type>, <type><payload>), ...] -> [(<type>, <payload>), ...]