Update gpg_sign_object full keyid workaround
We used to require the caller to supply a gpg keyid, to identify
the signing key's full keyid (via pubkey export), when creating a
signature with versions of gpg that don't add full keyids to
signatures.

This approach is not only inconvenient but also error prone when
dealing with key bundles. Since a caller could e.g. pass the master
key's keyid, while gpg actually signs with a subkey, which would
entail storing the wrong keyid (passed master) for the signature
(automatically used sub).

This commit changes the signing function to make use of the short
keyid, which per rfc4880 section 5.2.3.1 should always be on the
signature, when the full keyid is not present.

With the help of the short keyid we export a public key bundle, i.e
all associated master and subkeys and filter out the full keyid
with the matching suffix, which we then add to the signature.

This change also makes the unit test case handling obsolete that
suppressed signing tests with defaulft keyid for older versions
of gpg (travis). These case handlings are also removed by this commit.