Fix bug in TestInTotoVerifyThresholdsGpgSubkeys
The 4 test scenarios that include signing with a master key,
did not work as expected, since there is no way (known to me)
to sign with a gpg master key, if the key has a signing subkey.

This commit changes the MMM (see TestCase docstring) scenario to
use a key without subkeys and the MMS, MSM and MSS scenarios
to just assert for a mismatch of passed key and signing key.

Since gpg automatically uses a subkey if available, the scenarios
as they were, would translate to scenarios where a subkey is
used intentionally, like so:
MMM -> SMM
MMS -> SMS
MSM -> SSM
MSS -> SSS