FIX:test_verifylib: sort artifact queues
During the test of the modify rules, it is possible that the returned
queues are in different order than they came in (this is due to the
converstion to a set and back).

Sort the queues before comparing the two queues to ensure the contents
are compared, and not their order. Likewise, declare the elements in artifact
queues alfabetically to make sure the comparison stands.