Fix stdout to devnull in runtests subprocess call
In python 2.7 subprocess does not have the DEVNULL constant.
A cross version fix is to pass os.devnull (file) instead.

This commit also replaces the blanket except with a specific
exception. A comment above says that we can't use specific
except because the call in the try block raises one of two
exceptions. Why not catch both? Let's see if windows tests
complain.