Use custom logging.StreamHandler for testing
We used to suppress logging by replacing in_toto's base logger's
StreamHandler with a NullHandler for tests. While this eliminated
test log inundation, it was not helpful to troubleshoot failing
tests.

This commit implements a custom streamhandler for tests, in order
to allow unittest stdout/stderr buffering and only show output in
case a test fail.

Inspired by https://stackoverflow.com/a/33271004