Fix illegal modification of artifact queue
During artifact rule verification `verify_item_rules` may pass the
`source_artifacts_queue` to `verify_match_rule`, which returns an
updated version of the queue (minus matched artifacts) and
overrides the original `source_artifacts_queue`.

However, the queue is passed to `verify_match_rule` by reference,
and modified directly. While this wouldn't be a problem for
`source_artifacts_queue`, as it is modified and then returned to
override itself, there is another reference pointing to that list,
i.e. (depending on whether `verify_item_rules` operates on material
or product rules) `source_products_queue` or `source_artifacts_queue`,
neither of which should not be modified by `verify_match_rule`.

This commit changes `verify_match_rule` to not modify the passed
reference and also makes an actual copy of `source_materials_queue`
and `source_products_queue` in `verify_item_rules` just to be safe.