Refactor artifact rule verification (WIP) [ci skip]
It turns out that there are more problems in our
artifact rule verification than addressed in
4b3028aab6b83823b1c8c78b7a8e4d9244b5e554

The biggest problem is probably that it is very convoluted, i.e.
there are different "queues" that share the same lists in memory
and get modified under different conditions, which inevitably
leads (lead) to bugs.

This commit trys to align more closely with the pseudo-code
in the spec (which OTOH is a little bit too oversimplified, e.g.
there is only one queue, which IMHO is not enough).

TODO:
 - Try to simplify even more
 - Update docstrings
 - Fix tests / add better tests