Remove gpg signature key_expire_time field
Remove redundant parsing of key expiration time subpacket in
gpgp.common.parse_signature_packet. This and other subpackets are
also added to  the signature's opaque "info" field. Which can
be used for further intermediary processing of signature data
that does not need to end up in the ultimate signature.

This commit also changes the "info"'s "subpackets" list to a
dictionary, to consider the last subpacket of a type only, in case
there are more. Hashed subpackets are prioritized over unhashed.
This is the same subpacket ambiguity resolution scheme we use for
issuer keyid and fingerpint.