Re-add gpg verbosity in runtests check_usable_gpg
844366ac5cf41ef7665f4df229ac4616d092a9de and
0a63a0fa1191033cce5f97a1062df1e052f1394a modified the test
aggregators gpg availablity check to send the output to devnull,
with the goal to remove noise in the testlog.

However, knowing the available version is also helpful for
debugging issues on e.g. CI systems.

The function now also checks for gpg2 and only sets TEST_SKIP_GPG
if neither gpg2 nor gpg are available.

This commit also updates the caught error to type OSError,
which encompasses the possible errors cross-platform/version.
- OSError (on Python2)
- FileNotFoundError (on Python 3)
- WindowsError (on windows)