Fix pylint E1120 (no-value-for-parameter) in tests
Pylint revealed that TestUtil.test_unrecognized_key_type used to
assert for a test programming error, which is not very useful.

This commit changes the test to what I think the author might have
intended and removes a now obsolete `pragma: no cover`, that has
masqueraded the bogus test, and also fixes a bug in the docs.

A more in-depth revision shall be performed with
in-toto/in-toto#80.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>