Fix pylint (py2) import errors in tests
The errors were all related to importing mock as 3rd-party
backport on Py2 and as part of standard lib's unittest package
on Py3.

This commit also uses the same case handling for different versions
of mock in all relevant tests now, and updates ci requirements to
only install the backport if necessary.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>