Make pynacl optional
install_requires should list "what a project minimally needs to run
correctly"[1] and the presence of these dependencies is checked at runtime
when the project is installed via pip.

The securesystemslib pynacl extras allow for faster implementations of
cryptographic algorithms, but aren't required in order to run - therefore
they should not be listed in install_requires.

Unfortunately pypa/pip#4957 makes including the securesystemslib[pynacl] extra
in extras_require a noop when securesystemslib is already installed, therefore
we add pynacl itself to extras_require so that we can keep in-synch with the
pynacl dependency in securesystemslib

1. https://packaging.python.org/discussions/install-requires-vs-requirements

Signed-off-by: Joshua Lock <jlock@vmware.com>