Revise pylintrc
Prior to this commit our pylintrc was a full copy from
https://github.com/PyCQA/pylint/blob/master/pylintrc, listing
all defaults explicitly and adding a few customizations.

IMHO, the file is easier to read and maintain if it only contains
the customizations that we require.

This commit removes all explicit defaults and enables a few
checks that had been disabled intransparently, e.g. by their
category initials such as "C" and "R". Newly enabled checks
include:

- whitespace around operators
- line length (<80)
- consider using `in` (instead of e.g. a == "A" or a == "B")
- simplifiable if statement
- `ininstance` instead of `type ==`
- singleton comparison (`is` instead of `==`)
- naming conventions (uppercase constants, snake_case)
- missing docstrings

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>