Invoke tests explicitly when building for debian
The default way of invoking tests (dh_auto_test) during a debian
build `python setup.py test`.
This call also tries to resolve dependencies defined in setup.py,

As per setup.py in-toto depends on secureystemslib["crypto"], which
in turn depends on colorama>=0.3.9, which is not available in
Debian.

But in-toto and securesystemslib do work with a lower version of
colorama as well, which we install explicitly per in-toto's
debian/control, hence we can skip the setup.py-based dependency
resolving and just run the tests.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>