Fix tests that call python interpreter for debian
A few tests call the python interpreter in a subprocess, e.g. to
test or custom `process` module or to run simple in-toto step or
inspection commands.

In debian, where we only build for python3, "python" is not
available during build, and thus makes those tests fail.

This commit adds quilt downstream patches to change the used
interpreter to "python3" in the relevant tests (this approach is
backed by our Debian-man @h01ger).

Here's how I applied the patches and built in-toto 0.4.0 with
tests:

```
 # Assuming chroot is set up
 # See https://wiki.debian.org/sbuild

 # Download upstream tarball
wget https://github.com/in-toto/in-toto/releases/download/v0.4.0/in-toto-0.4.0.tar.gz \
  -O python-in-toto_0.4.0.orig.tar.gz

 # Extract upstream tarball
tar xf python-in-toto_0.4.0.orig.tar.gz

 # Clone upstream repo for debian files and tests
git clone https://github.com/in-toto/in-toto.git in-toto-upstream

 # Copy debian files
cp -r in-toto-upstream/debian in-toto-0.4.0/

 # Change into unpacked tarball
cd in-toto-0.4.0

 # Apply patch in debian/patches/
quilt push

 # Build with tests
sbuild -A -s -d unstable
```

Note, an alternative for most tests would be, to use
`sys.executable` instead of the hardcoded interpreter name. But
this does not work for the inspection ("python ./scripts/tar ...")
defined as string in "demo.layout.template".

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>