Add custom sphinx argparse-example-epilog extension
The in_toto command line tools use argparse epilog to add an
example usage section to the help message. The epilog does not
render nicely when using the sphinx-argparse extension.

This commit adds a custom extension to properly render the title
(expected to be 'EXAMPLE USAGE') and code blocks of a command line
tools epilog.

See ArgparseUsageEpilog docstring for details about the extension.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>