Manually check mutual exclusion in in-toto-record
Switch to manual mutual exclusion check for --key and --gpg
arguments (as in in_toto_run).

The `add_mutually_exclusive_group` used for the --key and --gpg
options can not be called on a custom argument group, such as the
"required named arguments" group. As a consequence the options
are listed in the "optional arguments" section. Since at least one
of them is required it is better to list both of them in the
"required named arguments" group (in addition to clarifying their
use in via the per-argument description.

This commit removes the use of add_mutually_exclusive_group and
siwtches to manual mutual exclusion check for --key and --gpg
arguments (as in in_toto_run) for better help message/cli docs.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>