Check type and not instance in test_common_args
When testing the undocumented (protected) attribute
'_action_groups' from 'argparse', it is safer to do a type and not
an instance check, because in 'common_args.sort_action_groups' we
replace what '_action_groups' references with a list. So if argparse
were to change its type to something list-ish, it could lead to
unexpected behavior on our side and we might not detect it in the
test here.

Kudos to @SantiagoTorres for pointing this out.

Pylint claims that this is an unidiomatic-typecheck. We beg to
differ.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>
Signed-off-by: cindykimxp <cindykim100@gmail.com>