Add windows dev mode hint to symlink tests
Added an extra method '_raise_win_dev_mode_error' to the 'Test_ApplyExcl
udePatterns' class in test_runlib.py, and updated 3 other methods to inc
lude it . It checks if the user is on a Windows platform. If so, they ar
e informed that Developer mode is required to work with symlinks on Wind
ows, and asked if it is enabled.

Signed-off-by: Isha Dave <ishavdave@nyu.edu>