Address pylint raise-missing-from warning
Versions 2.6.0 and later of pylint adhere to PEP 3134
and trigger a 'raise-missing-from' warning (W0707) when
chained exceptions are raised implicitly.

The 'from' keyword is a Python3.x feature, that is why
six.raise_from is used for Python2.x compatibility.

Kudos to @sechkova and @jku for showing how to do this
(theupdateframework/tuf#1116).

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>