Adopt recent sslib key interface changes
PR #402 adopted key interface changes from the pending
secure-systems-lab/securesystemslib#288 PR and was merged
prematurely. The sslib PR now has further evolved, in order to
follow the principle of secure defaults in regards to private key
encryption, which requires the following adoptions in in-toto:

- The not secure by default generate_and_write_*_keypair function
  is now protected (_generate_and_write_*_keypair), and only used
  for the keygen cli utility, where it is really convenient.

- In other cases we use either generate_and_write_*_keypair (for
  encrypted keys only) or generate_and_write_unencrypted_*_keypair.

Furthermore, the newly added sslib key generation interface
functions are added to the in-toto API docs.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>