Allow running in_toto_verify with prerecorded inspection results
Allow to verify the layout using pre-recorded inspection results. In
practice, this is very useful with a dummy steps ("/usr/bin/true"
command) when integrating with package managers (apt, dnf, etc) when the
file hashes are already known from repository metadata. This allows to
verify the packages based on links produced by rebuilders, before
actually downloading the packages (do not waste bandwidth for invalid
packages). The package manager will later verify if downloaded
packages actually matches expected hashes. This is especially useful for
DNF, where rejecting a package based on its hash (before downloading
it), is relatively easy, while rejecting already downloaded package is
much harder and prone to accidentally re-use previously rejected
package.

Implement this as an extra argument ("prerecorded_inspections") to the
in_toto_verify() function, which consists of materials and products
hashes for each step. Steps execution included in this dict will be
skipped and materials/products hashes from the argument will be used
instead.

Note not all the link parameters can be set in prerecorded_inspections
dict, but all that are set in run_all_inspections().

Signed-off-by: Marek Marczykowski-Górecki <marmarek@invisiblethingslab.com>