Case-handle cyptography test dependency for py2
- On Python3 test against the latest version.
- On Python2 test against the latest version supported on Python 2.

NOTE: This is only relevant for requirements-pinned.txt, where we
pin versions auto-bumped by Dependabot to trigger tests against
updates of direct and transitive dependencies.

When installing in-toto's dependencies via setup.py (or
requirements.txt) the "python_requires" field of the corresponding
dependency's setup.py makes sure that a supported version is
installed.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>