Add DSSE envelope convertor in Signable
As Link and Layout objects are Type Signable, it is suitable for a
DSSE envelope serializer or convertor.

The method "create_envelope" in Signable will serialize the Link or
Layout object into a payload of a payload_type (in-toto specific)
that is signable_bytes for now. A DSSE Envelope will be returned in
the end containing payload and payload_type.

Why signable_bytes would be good for payload?
- signable_bytes is canonical JSON byte sequence of Link or Layout
  that is currently used to create and verify signatures. In DSSE
  payload will be used to create and verify DSSE signatures.
- There is no other option to generate payload.

Signed-off-by: Pradyumna Krishna <git@onpy.in>