Change signing method in Metablock.sign
Metablock.sign is modified and now uses Signer to create signature.
This method will help in reducing duplication in adding support for
DSSE with current metadata wrapper Metablock.

Old signing method that uses key dict still exist and moved to a
different method with name sign_key. All references to the method
are changed, so it won't bring a breaking change.

Signed-off-by: Pradyumna Krishna <git@onpy.in>