Add SerializationMixin in Metablock metadata
SerializationMixin provides methods to serialize the data using
different serializers like `JSONSerializer` and deserialize them
using deserializers. These methods includes conversion of object
to and from bytes, and to and from file.

The SerializationMixin provide more flexibility to choose a
serializer for serialization and storage backend to store the
files.

Addition of this Mixin will result in a common way to dump and
load the metadata in in-toto using different wrappers, which will
reduce duplication in the code.

Signed-off-by: Pradyumna Krishna <git@onpy.in>