Generate metadata using DSSE in in_toto_run
This is the first step in order to add DSSE support to in-toto, A
use_dsse parameter is added to in_toto_run which selects the sign-
ature wrapper and generates metadata using that.

Metablock wrapper supports Signer to sign the signature and uses
SerializationMixin for serializaiton purporse, which is common for
both signature wrappers. This reduces duplication in the
implementation of DSSE envelope for in-toto.

To generate metadata with desired method:
- use_dsse parameter selects and creates an empty envelope.
- Signer is created according to the parameters provided and used
  to sign the envelope.
- In the end the envelope is dumped using to_file method present
  in signature wrappers.

Signed-off-by: Pradyumna Krishna <git@onpy.in>