Refactor AnyMetadata methods, add them to Metablock
AnyMetadata methods utilizes conditions to provide common
interfaces for some methods, which is not a good pattern to
follow. Hence, It is decided to Refactor them and add them to
Metablock and Envelope.

Refactoring Envelope and Metablock two methods:
* `deserialize_payload`: This method will return signed object
  after matching its type.
* `match_paylaod`: Matches type of the signed object.

`verify_signatures` method is not added to Metablock because
of current incompatibility of Metablock with Signer, Key and
Signature objects. It requires a major refactor in order to add
these methods to the Metablock.

AnyMetadata is reverted back to MetadataLoader.

Signed-off-by: Pradyumna Krishna <git@onpy.in>