Refactor Envelope with in-toto's Envelope
in-toto's Envelope is a subclass of the DSSE Envelope present at
securesystemslib with the in-toto payload deserializer.
For creation and verification of singature, similar interface for
Metablock is added.

The references to previous Envelope are changed, and other methods
like payload extraction, signing and verification etc.
Signed.create_envelope still create SSlibEnvelope, if it creates
a in-toto copy of Envtope then it will cause a circular import.

Some tests are fixed, public_key dict conversion to Key are added
wherever required for signature verification without any breaking
change.

Signed-off-by: Pradyumna Krishna <git@onpy.in>