JSON payload for DSSE and refactor create_envelope
Canonical JSON is not a type/subset of JSON and hence encoding
and decoding of canonical json works differently. Which creates
an problem in decoding canonical json, as we can't use
JSONDeserializer to deserializer it.

Using JSON for DSSE payload is favourable in this situation
because we don't have to create new deserializer and supports
both serialization and deserialization from and to the Signable
object.

create_envelope is refactored to from_signable method of in_toto
Envelope.

Signed-off-by: Pradyumna Krishna <git@onpy.in>