tests: adopt replacement of HAVE_GPG constant
secure-systems-lab/securesystemslib#437 replaced the HAVE_GPG
bool constant with a function with similar effect, to allow lazy
evaluation of whether GPG is present.

That is, we only shell out to check if gpg is present, when we
call have_gpg() and not when we import the module, where the name
is defined, which was the case before.

This commit adopts the rename in in-toto tests. No other code in
in-toto is affected by the rename.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>