Add FileResolver and use in runlib
In preparation for generic artifact URIs support, this commit
factors out ``record_artifacts_as_dict`` implementation to a
``hash_artifacts`` method of a new ``FileResolver`` class.

All configuration is done in the constructor, so that
hash_artifacts can be used as generic interface.

The code is simplified and modernized, while keeping strict
backwards-compatibility. This allows runlib tests to pass without
changes, except for removing an obsolete helper test, and adding a
bridge for another.

Code-style for the newly added resolver subpackage is black+isort,
and thus requires separate linting. This will be consolidated with
in-toto/in-toto#580.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>