runlib: add basic local material checker
Adds new runlib functionary API, to perform basic vetting of materials.
More specifically, to check if the **untrusted** materials match the
products of a passed **trusted** link metadata file.

**Caveat**
The used link metadata is not validated against a layout. This means, no
checking if the link is part of a certain supply chain, or if it was
signed by a threshold of authorized functionaries.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>